+++ dags/spark/ingestion_from_landing_data_file_to_bronze_tables_dag.py
@@ -13,13 +13,14 @@ from __future__ import annotations
 
 from datetime import timedelta
 
+from airflow.decorators import dag
+from airflow.providers.cncf.kubernetes.operators.spark_kubernetes import SparkKubernetesOperator
+from airflow.providers.cncf.kubernetes.sensors.spark_kubernetes import SparkKubernetesSensor
+from airflow.utils.dates import days_ago
+
 # [START import_module]
 # O objeto DAG; precisaremos disso para instanciar um DAG
-from airflow.decorators import dag
-
 # Operadores, precisamos que isso funcione!
-from airflow.providers.cncf.kubernetes.operators.spark_kubernetes import SparkKubernetesOperator
-from airflow.utils.dates import days_ago
 
 # [END import_module]
 
@@ -31,7 +32,7 @@ from airflow.utils.dates import days_ago
 # [FIM import_module]
 # Documentação baseada em Markdown que serão renderizados nas páginas Grid , Graph e Calendar.
 doc_md_DAG = """
-# DAG Entrega dos dados que vem de um csv para uma tabela iceberg no minio
+# DAG Entrega dos dados que vem da landing para uma tabela bronze
 
 Este é um exemplo de DAG que usa SparkKubernetesOperator e SparkKubernetesSensor.
 Neste exemplo, crio duas tarefas que são executadas sequencialmente.
@@ -40,8 +41,8 @@ E a segunda tarefa é verificar o estado final do sparkApplication que enviou no
 
 ## Objetivo desta DAG
 
-* Processar todos os dados da raw zone referentes aos dados de torre e passar para uma tabela
-iceberg no minio
+* Processar todos os dados da landing zone referentes aos dados de user,subscription,movies
+e credit_card, passando para uma tabela na camada bronze no minio
 
 Execute para testar.
 """
@@ -57,27 +58,27 @@ default_args = {
     "email_on_retry": False,
     "retries": 0,
     "retry_delay": timedelta(1),
-    "pool": "slow_pool",
 }
 # [FIM default_args]
 
 
 # [INICIO dag]
 @dag(
-    dag_id="bronze-to-silver",
+    dag_id="ingestion-from-landing-data-file-to-bronze-tables",
     default_args=default_args,
     start_date=days_ago(1),
     catchup=False,
     schedule_interval="@daily",
     max_active_runs=1,
-    tags=["spark", "kubernetes", "sensor", "iceberg", "minio", "s3", "bronze", "silver"],
+    tags=["spark", "kubernetes", "sensor", "iceberg", "minio", "s3", "raw", "bronze"],
     doc_md=doc_md_DAG,
 )
-def bronze_to_silver_dag() -> None:
+def ingestion_from_landing_data_file_to_bronze_tables_dag() -> None:
     """
-    `bronze_to_silver_dag()` é uma função que define um DAG
+    `ingestion_from_landing_data_file_to_bronze_tables_dag()` é uma função que define um DAG
     (Directed Gráfico acíclico) no Apache Airflow. Este DAG é responsável por ingerir
-    dados de uma tabela bronze para uma tabela silver. Consiste em duas tarefas:
+    dados dda landing, processar e colocar em uma tabela delta na camada bronze.
+    Consiste em duas tarefas:
     """
 
     # [INICIO set_tasks]
@@ -88,22 +89,22 @@ def bronze_to_silver_dag() -> None:
     # de yaml para acionar o processo, usando o spark-on-k8s para operar com base nos
     # dados e criando um `SparkApplication` em contêiner.
     submit = SparkKubernetesOperator(
-        task_id="bronze_to_silver_submit",
+        task_id="ingestion_from_landing_data_file_to_bronze_tables_submit",
         namespace="processing",
-        application_file="bronze_to_silver.yaml",
+        application_file="yamls/ingestion_from_landing_data_file_to_bronze_tables.yaml",
+        kubernetes_conn_id="conn_kubernetes",
         do_xcom_push=True,
         # O parâmetro `params` no `SparkKubernetesOperator` é usado para passar parâmetros
         # adicionais para o `SparkApplication` que será executado no cluster Kubernetes.
         # Esses parâmetros podem ser acessados no código do aplicativo Spark.
-        queue="kubernetes",
         params={
             "spark_driver_cores": 2,
             "spark_driver_memory": "2G",
             "spark_executor_cores": 2,
             "spark_executor_instances": 1,
             "spark_executor_memory": "2G",
-            "spark_job_name": "bronze-to-silver",
-            "spark_file": "bronze_to_silver.py",
+            "spark_job_name": "ingestion-from-landing-data-file-to-bronze-tables",
+            "spark_file": "spark/jobs/ingestion_from_landing_data_file_to_bronze_tables.py",
         },
         doc_md="""
         ### Proposta desta tarefa
@@ -114,18 +115,45 @@ def bronze_to_silver_dag() -> None:
         dados e criando um `SparkApplication` em contêiner.
         """,
     )
+    # A variável(task) `sensor` está criando uma instância da classe
+    # `SparkKubernetesSensor`. Este sensor é responsável por monitorar o status de um
+    # `SparkApplication` em execução em um cluster Kubernetes. Usando o sensor para ler
+    # e visualizar o resultado do `SparkApplication`, lê do xcom e verifica o par de
+    # status [chave e valor] do `submit`, contenco o nome do `SparkApplication` e
+    # passando para o `SparkKubernetesSensor`.
+    sensor = SparkKubernetesSensor(
+        task_id="ingestion_from_landing_data_file_to_bronze_tables_sensor",
+        namespace="processing",
+        application_name="{{task_instance.xcom_pull(task_ids='ingestion_from_landing_data_file_to_bronze_tables_submit')['metadata']['name']}}",  # noqa: E501
+        kubernetes_conn_id="conn_kubernetes",
+        attach_log=True,
+        doc_md="""
+        ### Proposta desta tarefa
+
+        * Ser responsável por monitorar o status de um `SparkApplication` em execução em um cluster
+        Kubernetes.
+
+        * Usar o sensor para ler e visualizar o resultado do `SparkApplication`.
+
+        * Ler do xcom e verifica o par de status [chave e valor] do `submit`, contenco o nome do
+        `SparkApplication` e passando para o `SparkKubernetesSensor`.
+        """,
+    )
     # [FIM set_tasks]
 
     # [INICIO task_sequence]
-    submit
+    # `submit >> sensor` está definindo a dependência entre a tarefa `submit` e a
+    # tarefa `sensor`. Isso significa que a tarefa `sensor` só começará a ser executada
+    # após a tarefa `submit` for concluída com sucesso.
+    submit >> sensor
     # [FIM task_sequence]
 
 
 # [FIM dag]
 
 # [INICIO start_dag]
-# `bronze_to_silver_dag()` está criando uma instância da DAG
-# `delivery-data-from-sap-hana-to-kafka`. Esta função(instância) pode ser usada para
+# `ingestion_from_landing_data_file_to_bronze_tables_dag()` está criando uma instância da DAG
+# `ingestion-from-landing-data-file-to-bronze-tables`. Esta função(instância) pode ser usada para
 # iniciar a execução da DAG no Apache Airflow.
-bronze_to_silver_dag()
+ingestion_from_landing_data_file_to_bronze_tables_dag()
 # [FIM start_dag]
