+++ dags/spark/yamls/ingestion_from_landing_data_file_to_bronze_tables.yaml
@@ -29,7 +29,7 @@ spec:
   mode: cluster
   image: gersonrs/spark-py:v1.0.0
   imagePullPolicy: IfNotPresent
-  mainApplicationFile: "local:///repo/airflow_dags.git/dags/spark_files/{{ params.spark_file }}"
+  mainApplicationFile: "local:///repo/airflow_dags.git/dags/{{ params.spark_file }}"
   sparkVersion: 3.5.0
   timeToLiveSeconds: 360
   restartPolicy:
@@ -55,7 +55,7 @@ spec:
     # - name: xcom
     #   emptyDir: {}
   driver:
-    cores: {{ params.spark_driver_cores }}
+    cores: "{{ params.spark_driver_cores }}"
     memory: "{{ params.spark_driver_memory }}"
     labels:
       version: 3.5.0
@@ -101,8 +101,8 @@ spec:
     #         cpu: "1m"
     #         memory: "10Mi"
   executor:
-    cores: {{ params.spark_executor_cores }}
-    instances: {{ params.spark_executor_instances }}
+    cores: "{{ params.spark_executor_cores }}"
+    instances: "{{ params.spark_executor_instances }}"
     memory: "{{ params.spark_executor_memory }}"
     labels:
       version: 3.5.0
