+++ dags/spark/yamls/raw_to_bronze.yaml
@@ -1,3 +1,21 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+---
 apiVersion: "sparkoperator.k8s.io/v1beta2"
 kind: SparkApplication
 metadata:
@@ -43,7 +61,7 @@ spec:
     - name: xcom
       emptyDir: {}
   driver:
-    cores: {{ params.spark_driver_cores }}
+    cores: "{{ params.spark_driver_cores }}"
     memory: "{{ params.spark_driver_memory }}"
     labels:
       version: 3.4.3
@@ -80,7 +98,8 @@ spec:
     sidecars:
       - name: airflow-xcom-sidecar
         image: alpine
-        command: ["sh", "-c", 'trap "exit 0" INT; while true; do sleep 1; done;']
+        command:
+          ["sh", "-c", 'trap "exit 0" INT; while true; do sleep 1; done;']
         volumeMounts:
           - name: xcom
             mountPath: /airflow/xcom
@@ -89,8 +108,8 @@ spec:
             cpu: "1m"
             memory: "10Mi"
   executor:
-    cores: {{ params.spark_executor_cores }}
-    instances: {{ params.spark_executor_instances }}
+    cores: "{{ params.spark_executor_cores }}"
+    instances: "{{ params.spark_executor_instances }}"
     memory: "{{ params.spark_executor_memory }}"
     labels:
       version: 3.4.3
