+++ dags/spark/yamls/transform_and_enrichment_from_bronze_to_silver.yaml
@@ -1,14 +1,14 @@
 apiVersion: "sparkoperator.k8s.io/v1beta2"
 kind: SparkApplication
 metadata:
-  name: delivery-data-from-silver-to-gold
+  name: transformation-and-enrichment-from-bronze-to-silver
   namespace: processing
 spec:
   type: Python
   mode: cluster
   image: "gersonrs/spark-py:v3.3.2"
   imagePullPolicy: IfNotPresent
-  mainApplicationFile: "local:///app/silver_to_gold.py"
+  mainApplicationFile: "local:///app/bronze_to_silver.py"
   sparkVersion: "3.3.2"
   restartPolicy:
     type: OnFailure
